backend/requirements.txt
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
click==8.1.8
fastapi==0.103.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.13.0
uvicorn==0.23.2
